Automate cert rotation & update credential jobs to use credhub
Adds a rotate-certs job that queries certs stored in credhub and rotates certs
due to expire in a month. For CAs, it follows
github.com/pivotal-cf/credhub-release/blob/master/docs/ca-rotation.md

It updates the following credential jobs to use credhub:
- rotate-cloudfoundry-credentials
- set-smoke-test-credentials
- rotate-prometheus-credentials
- rotate-database-encryption-keys

It also updates the check-certificates job to use credhub and alert on 15 days.