Implements usable web view of PaaS marketplace
Running `cf marketplace` produces a pretty unreadable view of the services and
their associated plans. This app (under tools/marketplace/) shows the same data
but in a more usable and accessible format in the browser.

It is implemented as a simple web application, using the gin-gonic framework.

The pattern for writing HTTP handlers has been slightly altered compared to the
framework's documentation. Intead of a handler being of type
    `func(*gin.Context)`

handlers now have the type
    `func(context.ContextInterface) error`.

This change has been made to facilitate testing the handlers in isolation. An
adapter function `handlers.GinWrap` has been written to marshal the context
types in to the handler, and to pass any errors returned by the handler in to
the context instance.

When registering a handler, it must be wrapped using the adapter function.

```
ginRouter.GET("/", handlers.GinWrap(httpHandlers.Index))
```

There are a number of other small, but notable, additions to it:

1. The `github.com/foolin/goiew` view engine is used to provide support for a
master/layout file. This has been imported because the default view engine for
gin-gonic doesn't have such a feature easily.

2. Error-handling middleware, in the `middleware` package, is used to catch
common errors (such as `ErrNotFound`) and render consistent error pages.

3. The `partial` function has been added to the view engine to support
rendering a partial view with a paased-in data object. The
`github.com/foolin/goiew` has an `include` function out of the box, but it
inherits the data of the view doing the including.