Create custom UAA clients for tenants in different deployments
In some cases, we'll grant a tenant a custom UAA client. The client should only
exist in a specific environmen, and its secret is managed in by our credhub
instance.

The configuration of the clients is managed in
`manifests/cf-manifest/data/100-tenant-uaa-client-config.yml`, and a task in
the `generate-cf-config` stage of the pipeline generates an ops file for Bosh
based on that config, for the given deployment env.