Add a concourse job to rotate broker credentials
We want to rotate the broker passwords. This commit adds a new job
that will rotate the passwords in Credhub. When the main
`create-cloudfoundry` deploy pipeline is next run it will update the
brokers to use the new credentials and update Cloud Controller to use
those new credentials too.

This operaiton is not zero-downtime; our brokers do not support
multiple credentials at once. But for the cases when we want to rotate
it's OK to have brief downtime for creating or updating services.
Apps talking to existing services aren't affected.

This does not rotate every broker-related secret. The RDS Broker's
`master_password_seed` property is not rotated. Thankfully we don't
need to rotate that right now because there is no easy way to rotate
it.