Synchronise router instances from Bosh to Auto Scaling Group
Bosh's AWS CPI does not support the Auto Scaling Groups, but we want to use
them to synchronise ALB target groups.

To do this, we have a step after a Bosh deploy which compares the current
instances in the group and the active instances as reported by Bosh, and adds
and removes from the Auto Scaling Group as appropriate.